Rename copy submodule
Since this conflicts with one of the python builtins, it makes problems if you add this folder to the classpath directly (like what pycharm does)
